requirements.txt
moviepy==1.0.3
Pillow==10.4.0
pytube==15.0.0
Requests==2.32.3
